Stratum 1.5 with spy mode operationnal.
spy_stamps composed from userland so far, but everything seems to be working. Need to see perfs from here.
On ACPI, the "RTT" is now about 19 cycles on a 3GHz machine when timestamping through 4 syscalls.

Also prepared generic stamp structure to init the move to uni paradigm as it was needed by spy mode.

Plus some fixes here and there.