A quick and rough port of patches to FreeBSD CURRENT
It is not a functionnal set of patches. Many critical things have
changed in addition to the already tricky 32/64 bit code. Timestamp
formats, capture modes, new bpf header ... 

The timestamping make it tricky to timestamp out of the bpf.c code,
which is going to eliminate the last bits that made TSMODE_SYSCLOCK
and TSMODE_FAIRCOMPARE different. Based on the difference in the 
number of instructions, the error between the two modes is essentially
negligeable anyway.

I essentially commit this patch for me to be able to work on it in Amsterdam.