This is giving me some massive headaches.
Probably the 10th rewrite of this code, but commiting it to start fresh again. A
big tree has been hiding the forest.

Spent quite some time wondring why I was loosing precision on the radclock mode
from the kernel clock. I thought the problem came from phat since it is the one
involved in adding deltas. So was trying to push the resolution of the fixed
point arithmetic down. The problem was not there.

To hold the number of seconds of current year, we need about 32 bits. Which
leaves only 32 bits for the fractional part in a 64 bit int (fixed point
arithmetique). So a loss in resolution of the order of 200 pico sec. That was
good logic on my part so far. But with the kernel updating the fixed point
data, that is another story. The data is updated HZ times a second, which makes
a godzillion losses of resolution that do accumulate (we work with cumulative
deltas obviously). At poll period = 16sec, that is about 16 mus error ...
arrghhh.

Only solution is to increase the resolution of the fractional part, so I have
been moving the code to the use of bintime. Much cleaner, and will be valid
beyond 2038. I think I got most things right, but there are a few issues left.
The logic is now completely different, so I need a bit of sleep and more
thinking. But will get there.

Crazy, it is not that many lines of codes, but I have already spent 2 days
fixing, coding and tracking all these problems.