Introduce a new pthread mutex to watch over the consistency of the raw data bundle list on insert and free.
I thought insert_rdb_in_list() and free_and_cherry_pick() could work without
this lock, but it does not. I cannot really see why, must be missing something
obvious, but there is a race condition bbetween the insertion and free. Bug
occurs when hammering NTP control packets to the NTP server in use. A very rare
case, but we should protect ourselves from it.

Originally, I pushed a lock free approach for performance reason. In the normal
mode of operation, it works well since the processing thread is awaken when data
is available. A very loose form of thread synchronisation, but good enough
except in extrem cases.

I would like to get rid of this mutex one day, when I have some free time ... 

In the meantime, it is actually likely that the implicit loose synchronisation
among the threads will prevent them from spinning on the mutex. I haven't tried
to check if this is true or not, but likely based on my original design. I also
tried to minimize the protected area in the free section to make sure the insert
section has a chance to grab the mutex if it ever spin on it (that's the
performance critical section).

Well, I am defeated on this one for the time being but the war is not over yet.
This has been useful to fix a few other parts of the code and improve the JDEBUG
prints.