Fix segfault due to NTP control packets.
In the previous version, the buffers I used to mangle the Ethernet and SLL
headers may have ended up being of different sizes. I think I was trying to
optimize the memory footprint too much.

Sadly that was a bad idea if we have a succession of very small NTP control
packet and larger NTP packet. The big one assumes there is memory allocated for
it write, but there isn't after I mangled the buffers.

I made sure all buffers have the same fixed size to fix the problem.

I wonder if that could be the cause of the RDB chain list bug when used
lock-free ... there was quite a bit of memory stomping, which could explain the
mistery ... (ahah?).

Also, re-thought the size of the pcap snapshot length in light of new
development or future extensions. IPv6 may be a nightmare if we face extensions
header, will see.