Relax test on overlapping stamps when checking stamp sanity.
Now that we allow out of order packets, this test is not relevant anymore.
Changed it into a debug warning, and replace error returning one with something
a bit less strict.
If the NTP client is producing several stamps per-poll period because of
retransmission, each stamp will be passed to the algo.